% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdcFlu20172018-package.R
\docType{package}
\name{cdcFlu20172018}
\alias{cdcFlu20172018}
\alias{cdcFlu20172018-package}
\title{cdcFlu20172018 - Code and submissions for 2017-2018 CDC flu prediction contest}
\description{
Code and submissions for 2017-2018 CDC flu prediction contest
}
\author{
Evan Ray <elray@umass.edu>
}

