% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{download_and_preprocess_flu_data}
\alias{download_and_preprocess_flu_data}
\title{Download and preprocess the latest CDC flu data, both national and regional}
\usage{
download_and_preprocess_flu_data()
}
\value{
data frame with latest flu data, preprocessed
}
\description{
Download and preprocess the latest CDC flu data, both national and regional
}
