% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_num_MMWR_weeks_in_first_season_year}
\alias{get_num_MMWR_weeks_in_first_season_year}
\title{return integer that's either 52 or 53: number of weeks in the first year of
a given season.}
\usage{
get_num_MMWR_weeks_in_first_season_year(season)
}
\arguments{
\item{season}{season in the format "2014/2015"}
}
\description{
return integer that's either 52 or 53: number of weeks in the first year of
a given season.
}
\details{
requires MMWRweek package
}
