% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_num_MMWR_weeks_in_year}
\alias{get_num_MMWR_weeks_in_year}
\title{return integer that's either 52 or 53: number of MMWR weeks in the given year}
\usage{
get_num_MMWR_weeks_in_year(year)
}
\arguments{
\item{year}{year in the format "2014" -- can be character or numeric}
}
\description{
return integer that's either 52 or 53: number of MMWR weeks in the given year
}
\details{
requires non-exported function start_date from MMWRweek package
}
