% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_official_observed_seasonal_quantities}
\alias{get_official_observed_seasonal_quantities}
\title{Compute season onset, peak week, and peak incidence}
\usage{
get_official_observed_seasonal_quantities(data, season, incidence_var)
}
\arguments{
\item{data}{a data frame containing at minimum columns named season,
season_week and a column with some sort of incidence measure}

\item{season}{the season to look at}

\item{incidence_var}{a character string naming the variable in the data
argument containing a measure of incidence, or an integer index}
}
\value{
a list with four entries:
  1) observed_onset_week, either an integer between first_CDC_season_week
    and last_CDC_season_week (inclusive), or "none"
  2) observed_peak_week, an integer between first_CDC_season_week and
    last_CDC_season_week (inclusive)
  3) observed_peak_inc, a numeric with the maximum value of the specified
    incidence measure between first_CDC_season_week and last_CDC_season_week
  4) observed_peak_inc_bin, character name of incidence bin for peak incidence
}
\description{
Compute season onset, peak week, and peak incidence
}
