% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_one_kde_prediction_file.R
\name{make_one_kde_prediction_file}
\alias{make_one_kde_prediction_file}
\title{Wrapper for prediction of a single time-point from KDE fits}
\usage{
make_one_kde_prediction_file(path, season, season_week, n_sim)
}
\arguments{
\item{path}{filepath to fitted models}

\item{season}{character string of season for current fit, in format "XXXX/YYYY"}

\item{season_week}{numeric season week of first week for which predictions are desired}

\item{n_sim}{number of simulations to run for predictive distributions}
}
\value{
NULL just saves a file
}
\description{
Wrapper for prediction of a single time-point from KDE fits
}
