% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_predictions_plots}
\alias{make_predictions_plots}
\title{Make plots of prediction submissions for flu contest: so far, seasonal targets only}
\usage{
make_predictions_plots(preds_save_file, plots_save_file, data)
}
\arguments{
\item{preds_save_file}{path to a file with predictions in csv submission format}

\item{plots_save_file}{path to a pdf file where plots should go}

\item{data}{data observed so far this season}
}
\description{
Make plots of prediction submissions for flu contest: so far, seasonal targets only
}
